avoiding dependencies between saas_portal and saas_server, all features in saas_portal, saas_server will be just to collect information from clients.
