+++ saas_portal/models/__init__.py
@@ -1,2 +1,3 @@
 import res_config
 import saas_portal
+import wizard
