+++ saas_portal/models/saas_portal.py
@@ -3,6 +3,80 @@ import openerp
 from openerp import models, fields, api, SUPERUSER_ID
 from openerp.addons.saas_utils import connector, database
 from openerp import http
+from openerp.tools import config
+
+
+class SaasServerPlan(models.Model):
+    _name = 'saas_server.plan'
+
+    name = fields.Char('Plan')
+    template = fields.Char('Template')
+    demo = fields.Boolean('Demo Data')
+    sequence = fields.Integer('Sequence')
+    state = fields.Selection([('draft', 'Draft'), ('confirmed', 'Confirmed')],
+                             'State', default='draft')
+    role_id = fields.Many2one('saas_server.role', 'Role')
+    required_addons_ids = fields.Many2many('ir.module.module',
+                                           rel='company_required_addons_rel',
+                                           id1='company_id', id2='module_id',
+                                           string='Required Addons')
+    optional_addons_ids = fields.Many2many('ir.module.module',
+                                           rel='company_optional_addons_rel',
+                                           id1='company_id', id2='module_id',
+                                           string='Optional Addons')
+
+    _order = 'sequence'
+
+    def create_template(self, cr, uid, ids, context=None):
+        obj = self.browse(cr, uid, ids[0])
+        openerp.service.db.exp_create_database(obj.template, obj.demo, 'en_US')
+        addon_names = [x.name for x in obj.required_addons_ids]
+        if 'saas_client' not in addon_names:
+            addon_names.append('saas_client')
+        to_search = [('name', 'in', addon_names)]
+        addon_ids = connector.call(obj.template, 'ir.module.module',
+                                   'search', to_search)
+        for addon_id in addon_ids:
+            connector.call(obj.template, 'ir.module.module',
+                           'button_immediate_install', addon_id)
+        return self.write(cr, uid, obj.id, {'state': 'confirmed'})
+
+    def edit_template(self, cr, uid, ids, context=None):
+        obj = self.browse(cr, uid, ids[0])
+        d = config.get('local_url')
+        url = '%s/login?db=%s&login=admin&key=admin' % (d, obj.template)
+        return {
+            'type': 'ir.actions.act_url',
+            'target': 'self',
+            'name': 'Edit Template',
+            'url': url
+        }
+
+    def upgrade_template(self, cr, uid, ids, context=None):
+        obj = self.browse(cr, uid, ids[0])
+        return {
+            'type': 'ir.actions.act_window',
+            'view_type': 'form',
+            'view_mode': 'form',
+            'res_model': 'saas.config',
+            'target': 'new',
+            'context': {
+                'default_action': 'upgrade',
+                'default_database': obj.template
+            }
+        }
+
+    def delete_template(self, cr, uid, ids, context=None):
+        obj = self.browse(cr, uid, ids[0])
+        openerp.service.db.exp_drop(obj.template)
+        return self.write(cr, uid, obj.id, {'state': 'draft'})
+
+
+class SaasServerRole(models.Model):
+    _name = 'saas_server.role'
+
+    name = fields.Char('Name', size=64)
+    code = fields.Char('Code', size=64)
 
 
 class OauthApplication(models.Model):
@@ -70,78 +144,10 @@ class OauthApplication(models.Model):
             self.plan = access_token.user_id.plan_id.name
 
 
-class SaasConfig(models.TransientModel):
-    _name = 'saas.config'
+class ResUsers(models.Model):
+    _name = 'res.users'
+    _inherit = 'res.users'
 
-    action = fields.Selection([('edit', 'Edit'), ('upgrade', 'Upgrade')],
-                                'Action')
-    database = fields.Char('Database', size=128)
-    update_addons = fields.Char('Update Addons', size=256)
-    install_addons = fields.Char('Install Addons', size=256)
-    fix_ids = fields.One2many('saas.config.fix', 'config_id', 'Fixes')
-    description = fields.Text('Description')
-
-    def execute_action(self, cr, uid, ids, context=None):
-        res = False
-        obj = self.browse(cr, uid, ids[0], context)
-        method = '%s_database' % obj.action
-        if hasattr(self, method):
-            res = getattr(self, method)(cr, uid, obj, context)
-        return res
-
-    def edit_database(self, cr, uid, obj, context=None):
-        params = (obj.database.replace('_', '.'), obj.database)
-        url = 'http://%s/login?db=%s&login=admin&key=admin' % params
-        return {
-            'type': 'ir.actions.act_url',
-            'target': 'self',
-            'name': 'Edit Database',
-            'url': url
-        }
-
-    def upgrade_database(self, cr, uid, obj, context=None):
-        dbs = obj.database and [obj.database] or database.get_market_dbs()
-        uaddons = obj.update_addons and obj.update_addons.split(',') or []
-        update_domain = [('name', 'in', uaddons)]
-        iaddons = obj.install_addons and obj.install_addons.split(',') or []
-        install_domain = [('name', 'in', iaddons)]
-        no_update_dbs = []
-        for db_name in dbs:
-            try:
-                registry = openerp.modules.registry.RegistryManager.get(db_name)
-                with registry.cursor() as rcr:
-                    # update database.uuid
-                    openerp.service.db._drop_conn(rcr, db_name)
-                    module = registry['ir.module.module']
-                    # 1. Update existing modules
-                    uaids = module.search(rcr, SUPERUSER_ID, update_domain)
-                    if uaids:
-                        module.button_upgrade(rcr, SUPERUSER_ID, uaids)
-                    # 2. Install new modules
-                    iaids = module.search(rcr, SUPERUSER_ID, install_domain)
-                    if iaids:
-                        module.button_immediate_install(rcr, SUPERUSER_ID, iaids)
-                    # 3. Execute methods
-                    for fix in obj.fix_ids:
-                        getattr(registry[fix.model], fix.method)(rcr, SUPERUSER_ID)
-            except:
-                no_update_dbs.append(db_name)
-        if no_update_dbs:
-            desc = 'These databases were not updated: %s', ', '.join(no_update_dbs)
-            self.write(cr, uid, obj.id, {'description': desc})
-        return {
-            'type': 'ir.actions.act_window',
-            'view_type': 'form',
-            'view_mode': 'form',
-            'res_model': 'saas.config',
-            'res_id': obj.id,
-            'target': 'new',
-        }
-
-
-class SaasConfigFix(models.TransientModel):
-    _name = 'saas.config.fix'
-
-    model = fields.Char('Model', required=1, size=64)
-    method = fields.Char('Method', required=1, size=64)
-    config_id = fields.Many2one('saas.config', 'Config')
+    plan_id = fields.Many2one('saas_server.plan', 'Plan')
+    organization = fields.Char('Organization', size=64)
+    database = fields.Char('Database', size=64)
