+++ saas_server/__openerp__.py
@@ -4,11 +4,9 @@
     'author': 'Ivan Yelizariev',
     'category': 'SaaS',
     'website': 'https://it-projects.info',
-    'depends': ['auth_oauth', 'saas_portal', 'saas_utils'],
+    'depends': ['auth_oauth', 'saas_utils'],
     'data': [
-        'data/auth_oauth_data.xml',
-        'views/saas_server.xml',
-        'views/signup.xml'
+        'data/auth_oauth_data.xml'
     ],
     'installable': True,
     'description': '''
