+++ saas_server/models/saas_server.py
@@ -2,7 +2,6 @@ import os
 import openerp
 from openerp import SUPERUSER_ID
 from openerp import models, fields
-from openerp.tools import config
 from openerp.addons.saas_utils import connector, database
 
 
@@ -15,80 +14,6 @@ def get_size(start_path='.'):
     return total_size
 
 
-class SaasServerPlan(models.Model):
-    _name = 'saas_server.plan'
-
-    name = fields.Char('Plan')
-    template = fields.Char('Template')
-    demo = fields.Boolean('Demo Data')
-    sequence = fields.Integer('Sequence')
-    state = fields.Selection([('draft', 'Draft'), ('confirmed', 'Confirmed')],
-                             'State', default='draft')
-    role_id = fields.Many2one('saas_server.role', 'Role')
-    required_addons_ids = fields.Many2many('ir.module.module',
-                                           rel='company_required_addons_rel',
-                                           id1='company_id', id2='module_id',
-                                           string='Required Addons')
-    optional_addons_ids = fields.Many2many('ir.module.module',
-                                           rel='company_optional_addons_rel',
-                                           id1='company_id', id2='module_id',
-                                           string='Optional Addons')
-    client_ids = fields.One2many('saas_server.client', 'plan_id', 'Clients')
-
-    _order = 'sequence'
-
-    def create_template(self, cr, uid, ids, context=None):
-        obj = self.browse(cr, uid, ids[0])
-        openerp.service.db.exp_create_database(obj.template, obj.demo, 'en_US')
-        addon_names = [x.name for x in obj.required_addons_ids]
-        if 'saas_client' not in addon_names:
-            addon_names.append('saas_client')
-        to_search = [('name', 'in', addon_names)]
-        addon_ids = connector.call(obj.template, 'ir.module.module',
-                                   'search', to_search)
-        for addon_id in addon_ids:
-            connector.call(obj.template, 'ir.module.module',
-                           'button_immediate_install', addon_id)
-        return self.write(cr, uid, obj.id, {'state': 'confirmed'})
-
-    def edit_template(self, cr, uid, ids, context=None):
-        obj = self.browse(cr, uid, ids[0])
-        d = config.get('local_url')
-        url = '%s/login?db=%s&login=admin&key=admin' % (d, obj.template)
-        return {
-            'type': 'ir.actions.act_url',
-            'target': 'self',
-            'name': 'Edit Template',
-            'url': url
-        }
-
-    def upgrade_template(self, cr, uid, ids, context=None):
-        obj = self.browse(cr, uid, ids[0])
-        return {
-            'type': 'ir.actions.act_window',
-            'view_type': 'form',
-            'view_mode': 'form',
-            'res_model': 'saas.config',
-            'target': 'new',
-            'context': {
-                'default_action': 'upgrade',
-                'default_database': obj.template
-            }
-        }
-
-    def delete_template(self, cr, uid, ids, context=None):
-        obj = self.browse(cr, uid, ids[0])
-        openerp.service.db.exp_drop(obj.template)
-        return self.write(cr, uid, obj.id, {'state': 'draft'})
-
-
-class SaasServerRole(models.Model):
-    _name = 'saas_server.role'
-
-    name = fields.Char('Name', size=64)
-    code = fields.Char('Code', size=64)
-
-
 class SaasServerClient(models.Model):
     _name = 'saas_server.client'
 
@@ -97,7 +22,6 @@ class SaasServerClient(models.Model):
     users_len = fields.Integer('Count users')
     file_storage = fields.Integer('File storage (MB)')
     db_storage = fields.Integer('DB storage (MB)')
-    plan_id = fields.Many2one('saas_server.plan', 'Plan')
 
     def update_all(self, cr, uid, server_db):
         db_list = database.get_market_dbs(with_templates=False)
@@ -139,12 +63,3 @@ class SaasServerClient(models.Model):
                 res.append(data)
 
         return res
-
-
-class ResUsers(models.Model):
-    _name = 'res.users'
-    _inherit = 'res.users'
-
-    plan_id = fields.Many2one('saas_server.plan', 'Plan')
-    organization = fields.Char('Organization', size=64)
-    database = fields.Char('Database', size=64)
