--- saas_server/views/saas_server.xml
@@ -1,70 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<openerp>
-    <data>
-
-        <record id="view_plans_tree" model="ir.ui.view">
-            <field name="name">saas_server.plans.tree</field>
-            <field name="model">saas_server.plan</field>
-            <field name="priority">4</field>
-            <field name="arch" type="xml">
-                <tree string="Plans">
-					<field name="name"/>
-					<field name="template"/>
-					<field name="state" />
-					<button name="create_template" type="object" states="draft" icon="gtk-apply" />
-					<button name="edit_template" type="object" states="confirmed" icon="gtk-edit" />
-					<button name="upgrade_template" type="object" states="confirmed" icon="gtk-execute" />
-					<button name="delete_template" type="object" states="confirmed" icon="gtk-cancel" />
-                </tree>
-            </field>
-        </record>
-
-
-        <record id="view_plans_form" model="ir.ui.view">
-            <field name="name">saas_server.plans.form</field>
-            <field name="model">saas_server.plan</field>
-            <field name="arch" type="xml">
-				<form string="plans">
-					<group>
-						<field name="name"/>
-						<field name="template"/>
-						<field name="role_id"/>
-						<field name="demo"/>
-						<field name="sequence"/>
-					</group>
-					<notebook>
-						<page string="Required Addons" >
-							<field name="required_addons_ids" nolabel="1" />
-						</page>
-						<page string="Optional Addons" >
-							<field name="optional_addons_ids" nolabel="1" />
-						</page>
-					</notebook>
-				</form>
-            </field>
-        </record>
-
-        <record id="action_plans" model="ir.actions.act_window">
-            <field name="name">Plans</field>
-            <field name="type">ir.actions.act_window</field>
-            <field name="res_model">saas_server.plan</field>
-            <field name="view_type">form</field>
-            <field name="view_mode">tree,form</field>
-        </record>
-
-		<menuitem action="action_plans" id="menu_plans" parent="saas_portal.menu_saas" sequence="7"/>
-		
-		<record id="action_upgrade_clients" model="ir.actions.act_window">
-            <field name="name">Upgrade Clients</field>
-            <field name="type">ir.actions.act_window</field>
-            <field name="res_model">saas.config</field>
-            <field name="view_type">form</field>
-            <field name="view_mode">form</field>
-			<field name="target">new</field>
-			<field name="context">{'default_action': 'upgrade'}</field>
-        </record>
-
-		<menuitem action="action_upgrade_clients" id="menu_upgrade_clients" parent="saas_portal.menu_saas" sequence="8"/>
-
-    </data>
-</openerp>
